+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Fix async helpers not working when cache enabled
   * deps: handlebars@4.1.2
   * deps: walk@2.3.14